// src/app/shared/models/project.interface.ts
export interface Project {
    id: string;
    name?: string;
    coverImage?: any;
    title?: string;
    description?: string;
    longDescription?: string;
    technologies?: string[];  
    githubLink?: string;      
    month?: string;
    year?: string;
    code?: string;
    status?: string;
    tags?: string[];
}